requirements.txt
# ChainMind - HashKey Chain Horizon Hackathon
# Python 3.10+

# Web
flask>=3.0.0
flask-cors>=4.0.0

# Chain
web3>=6.0.0

# AI / Data
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Tasks
apscheduler>=3.10.0

# Config & HTTP
python-dotenv>=1.0.0
requests>=2.31.0

# Optional: lightweight LLM (uncomment if using)
# transformers>=4.30.0
# torch>=2.0.0
